+++ frontend/app/src/components/account-management/AccountManagementAside.vue
@@ -5,13 +5,7 @@ defineSlots<{
 </script>
 
 <template>
-  <div :class="$style.aside">
+  <div class="max-w-[31rem] mx-auto lg:max-w-[29%] flex-col grow-0 lg:flex-1 border-b lg:border-b-0 border-default lg:border-l overflow-auto">
     <slot />
   </div>
 </template>
-
-<style lang="scss" module>
-.aside {
-  @apply max-w-[31rem] mx-auto lg:max-w-[29%] flex-col flex-grow-0 lg:flex-1 border-b lg:border-l overflow-auto;
-}
-</style>