+++ frontend/app/src/components/app/AppHost.vue
@@ -2,8 +2,12 @@
 import { useBackendManagement } from '@/composables/backend';
 import { useSessionStateCleaner } from '@/composables/session/logout';
 import { useLocale } from '@/composables/session/use-locale';
+import { useThemeChecker } from '@/modules/theme/use-theme-checker';
 import { useSessionSettingsStore } from '@/store/settings/session';
 import { checkIfDevelopment, startPromise } from '@shared/utils';
+import '@/utils/chartjs-adapter-dayjs';
+
+useThemeChecker();
 
 const DevApp = defineAsyncComponent(() => import('@/DevApp.vue'));